move database.
