+++ .gitignore
@@ -6,6 +6,7 @@
 .pnp.js
 .yarn/install-state.gz
 secret.json
+newSecret.json
 # testing
 /coverage
